requirements.txt
bcrypt==3.1.2
cffi==1.9.1
Django==1.10.5

pycparser==2.17
python-dateutil==2.6.0
six==1.10.0
